Move extension to correct file
